backend/requirements.txt
# Backend requirements (ajout à requirements.txt)
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic-settings==2.6.1
python-multipart==0.0.12
sse-starlette==2.2.0
redis==5.2.0
rq==2.0.0
requests>=2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
